src/requirements.txt
scikit-learn 
pyodbc 
uvicorn[standard] 
python-multipart 
python-dotenv 
fastapi 
apscheduler